CWeaponMgr: Make use of size_t in IncrCount()
Keeps this function consistent with the previous change.

+++ Runtime/Weapon/CWeaponMgr.cpp
@@ -27,7 +27,7 @@ void CWeaponMgr::IncrCount(TUniqueId uid, EWeaponType type) {
   if (GetIndex(uid) < 0) {
     Add(uid, type);
   } else {
-    x0_weapons[uid][u32(type)]++;
+    x0_weapons[uid][size_t(type)]++;
   }
 }
